move factory into utils
